Always play indent beep when saying current line
When NVDA is configured to beep the indentation level, the beep is not
played if the line being spoken is at the same indent level as the
previously spoken line.

This is sensible, but does not behave well in interplay with explicit
commands to read the current line -- the indentation level of the
current line is not beeped (because it is the same as the "previously
read" line, i.e. the current line).

Make explicit read current line commands always beep the current
indentation level.